Ok, a lot happened in this commit...
Basically it comes down to integrating Unity and Socket.IO, which
is a bit finicky at times.

Most importantly I made it possible for players to join and leave
the game as it's being played. This involved a lot of moving stuff
around, including changing the scene and modifying the code for
the arena player and manager. I tried avoiding it but it was hard
to get stuff working without doing so - sorry!

Note that I found an okay way for sort-of-singletons that persist
between scenes :)